edit user

+++ app/Http/Controllers/UserController.php
@@ -7,6 +7,8 @@ use Illuminate\Http\Request;
 use App\User;
 use Illuminate\Support\Facades\View;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Input;
+use Illuminate\Http\Concerns\InteractsWithFlashData;
 
 class UserController extends Controller
 {
@@ -80,7 +82,14 @@ class UserController extends Controller
      */
     public function edit($id)
     {
-        //
+        $user=User::find($id);
+        $user_plans=DB::table('user_plans')
+            ->select('*')
+            ->whereRaw('user_id = :id',[ 'id' => $id] )
+            ->get();
+        $plans=Plan::all()->pluck('plan_name', 'id');
+        //return $plans;
+        return view('edituser')->with('user', $user)->with('user_plans',$user_plans)->with('plans',$plans);
     }
 
     /**
@@ -90,9 +99,33 @@ class UserController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request,$id)
     {
-        //
+        $user = User::find($id);
+        $user->firstname = Input::get('firstname');
+        $user->lastname = Input::get('lastname');
+        $user->email = Input::get('email');
+        $user->phone = Input::get('phone');
+        $user_plans=DB::table('user_plans')
+            ->select('*')
+            ->whereRaw('user_id = :id',[ 'id' => $id] )
+            ->delete();
+        $input_plans=Input::get('plans');
+        if(count($input_plans)>0){
+            foreach ($input_plans as $plan):
+                DB::table('user_plans')->insert(
+                    ['user_id' => $user->id, 'plan_id' => $plan]
+                );
+            endforeach;
+        }else{
+            DB::table('user_plans')->insert(
+                ['user_id' => $user->id, 'plan_id' => $input_plans]
+            );
+        }
+        $user->save();
+
+        $request->session()->flash('message', 'Successfully updated nerd!');
+        return redirect('users');
     }
 
     /**
@@ -105,6 +138,11 @@ class UserController extends Controller
     {
         $user=User::find($id);
         $user->delete();
+        DB::table('user_plans')
+            ->select('*')
+            ->whereRaw('user_id = :id',[ 'id' => $id] )
+            ->delete();
+
         return  redirect('users');
     }
 }
